% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexlm-package.R
\docType{package}
\name{complexlm-package}
\alias{complexlm}
\alias{complexlm-package}
\title{complexlm: Linear fitting for complex valued data}
\description{
Tools for linear fitting with complex variables. Includes ordinary least-squares and robust methods. Originally adapted from the rlm functions of MASS and the lm functions of stats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/QuantumOfMoose/complexlm}
}

}
\author{
\strong{Maintainer}: William Ryan \email{wryan1@binghamton.edu} (Adapted MASS rlm and stats lm for complex variables then spun that off into this package) [copyright holder]

Authors:
\itemize{
  \item Brian Ripley \email{ripley@stats.ox.ac.uk} [copyright holder]
}

Other contributors:
\itemize{
  \item Bill Venables [contributor]
  \item Douglas M. Bates [contributor]
  \item Kurt Hornik (partial port ca 1998) [translator]
  \item Albrecht Gebhardt (partial port ca 1998) [translator]
  \item David Firth [contributor]
  \item Bruce White [thesis advisor]
}

}
\keyword{internal}
