% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{cov}
\alias{cov}
\alias{cor}
\alias{var}
\title{Variance, Covariance, and Correlation for Complex Data}
\usage{
cov(x, y = NULL, na.rm = FALSE, method = "pearson", use, ...)

cor(x, y = NULL, na.rm = FALSE, use, method = "pearson", ...)

var(x, y = NULL, na.rm = FALSE, use, ...)
}
\arguments{
\item{x}{a numeric or complex vector, matrix, or dataframe.}

\item{y}{NULL (default) or a numeric vector, matrix, or dataframe with dimensions compatible with x.}

\item{na.rm}{logical. Should missing values be removed? Only considered when \code{x} is a vector.}

\item{use}{character string giving the desired method of computing covariances in the presence of missing values. Options are "everything" (default),
"all.obs", "complete.obs", or "na.or.complete". See \link[stats:cor]{stats::cov} for explanation of what each one does. Note that "pairwise.complete.obs" is not available for this complex method.}
}
\value{
numeric or complex the sample variance, covariance, or correlation of the input data.
}
\description{
Wrappers of \link[stats:cor]{stats::var}, \link[stats:cor]{stats::cov}, and \link[stats:cor]{stats::cor} that are capable of handling complex input.
}
\details{
For vector input, the sample variance is calculated as,\cr
\eqn{sum(Conj( mean(x) - x ) * ( mean(x) - x )) / (length(x) - 1)}\cr
And the sample covariance is calculated as, \cr
\eqn{sum(Conj( mean(x) - x ) * ( mean(y) - y )) / (length(x) - 1)}\cr
The Pearson correlation coefficient, which is the only kind available for complex data, is the covariance divided by the product of the standard deviations of all variables.
}
\section{Functions}{
\itemize{
\item \code{cor}: Correlation coefficient of complex variables.

\item \code{var}: S3 Variance of complex variables, a synonym for \link{cov}.
}}

\examples{
set.seed(4242)
n <- 9
foo <- complex(real = rnorm(n), imaginary = rnorm(n))
var(foo)
bar <- complex(real = rnorm(n), imaginary = rnorm(n))
var(x = foo, y = bar)
foobar <- data.frame(foo, bar)
cov(foobar)
cor(foobar)
}
