% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqs.R
\name{cov.rob}
\alias{cov.rob}
\alias{cov.mve}
\alias{cov.mcd}
\title{Resistant Estimation of Multivariate Location and Scatter, for Numeric or Complex Variables}
\usage{
cov.rob(
  x,
  cor = FALSE,
  quantile.used = floor((n + p + 1)/2),
  method = c("mve", "mcd", "classical"),
  nsamp = "best",
  seed
)

cov.mve(...)

cov.mcd(...)
}
\arguments{
\item{x}{
    a matrix or data frame.
  }

\item{cor}{
    should the returned result include a correlation matrix?
  }

\item{quantile.used}{
    the minimum number of the data points regarded as \code{good} points.
  }

\item{method}{
    the method to be used -- minimum volume ellipsoid, minimum
    covariance determinant or classical product-moment. Using
    \code{cov.mve} or \code{cov.mcd} forces \code{mve} or \code{mcd}
    respectively.
  }

\item{nsamp}{
    the number of samples or \code{"best"} or \code{"exact"} or
    \code{"sample"}.  The limit
    If \code{"sample"} the number chosen is \code{min(5*p, 3000)}, taken
    from Rousseeuw and Hubert (1997). If \code{"best"} exhaustive
    enumeration is done up to 5000 samples: if \code{"exact"}
    exhaustive enumeration will be attempted.
  }

\item{seed}{
    the seed to be used for random sampling: see \code{\link{RNGkind}}. The
    current value of \code{.Random.seed} will be preserved if it is set.
  }

\item{...}{arguments to \code{cov.rob} other than \code{method}.}
}
\value{
A list with components

  \item{center}{
    the final estimate of location.
  }
  \item{cov}{
    the final estimate of scatter.
  }
  \item{cor}{
    (only is \code{cor = TRUE}) the estimate of the correlation
    matrix.
  }
  \item{sing}{
    message giving number of singular samples out of total
  }
  \item{crit}{
    the value of the criterion on log scale. For MCD this is
    the determinant, and for MVE it is proportional to the volume.
  }
  \item{best}{
    the subset used. For MVE the best sample, for MCD the best
    set of size \code{quantile.used}.
  }
  \item{n.obs}{
    total number of observations.
}
}
\description{
Compute a multivariate location and scale estimate with a high
  breakdown point -- this can be thought of as estimating the mean and
  covariance of the \code{good} part of the data. \code{cov.mve} and
  \code{cov.mcd} are compatibility wrappers.
}
\details{
For method \code{"mve"}, an approximate search is made of a subset of
  size \code{quantile.used} with an enclosing ellipsoid of smallest volume; in
  method \code{"mcd"} it is the volume of the Gaussian confidence
  ellipsoid, equivalently the determinant of the classical covariance
  matrix, that is minimized. The mean of the subset provides a first
  estimate of the location, and the rescaled covariance matrix a first
  estimate of scatter. The Mahalanobis distances of all the points from
  the location estimate for this covariance matrix are calculated, and
  those points within the 97.5\% point under Gaussian assumptions are
  declared to be \code{good}. The final estimates are the mean and rescaled
  covariance of the \code{good} points.

  The rescaling is by the appropriate percentile under Gaussian data; in
  addition the first covariance matrix has an \emph{ad hoc} finite-sample
  correction given by Marazzi.

  For method \code{"mve"} the search is made over ellipsoids determined
  by the covariance matrix of \code{p} of the data points. For method
  \code{"mcd"} an additional improvement step suggested by Rousseeuw and
  van Driessen (1999) is used, in which once a subset of size
  \code{quantile.used} is selected, an ellipsoid based on its covariance
  is tested (as this will have no larger a determinant, and may be smaller).

  There is a hard limit on the allowed number of samples, \eqn{2^{31} -
  1}{2^31 - 1}.  However, practical limits are likely to be much lower
  and one might check the number of samples used for exhaustive
  enumeration, \code{combn(NROW(x), NCOL(x) + 1)}, before attempting it.
}
\section{Functions}{
\itemize{
\item \code{cov.mve}: comparability wrapper for \code{cov.rob}

\item \code{cov.mcd}: comparability wrapper for \code{cov.rob}
}}

\note{
This function relies on interquartile range, which orders the inputs as part of the calculation. The concept of ordering does not translate to the complex numbers,
so this function applies interquartile range to the real and imaginary components separately. The results of this will not be rotationally invariant, so this function
should be used with caution for complex variables.
}
\examples{
n <- 16
x <- matrix(rnorm(n), ncol = 2)
cov.rob(x)
}
