% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lqs.R
\name{lmsreg}
\alias{lmsreg}
\alias{ltsreg}
\title{Comparability Wrapper for lqs}
\usage{
lmsreg(...)

ltsreg(...)
}
\arguments{
\item{...}{arguments to be passed to \code{lqs.default} or
    \code{lqs.control}, see \code{control} above and \code{Details}.}
}
\description{
This is just a wrapper for \link{lqs}. See \link{lqs} for documentation.
}
\section{Functions}{
\itemize{
\item \code{ltsreg}: comparability wrapper for \code{lqs}
}}

