% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{mahalanobis}
\alias{mahalanobis}
\title{Mahalanobis Distance, with better complex behavior}
\usage{
mahalanobis(x, center, cov, inverted = FALSE, ...)
}
\arguments{
\item{x}{vector or matrix of data with, say, \eqn{p} columns.}

\item{center}{mean vector of the distribution or second data vector of
    length \eqn{p} or recyclable to that length.  If set to
    \code{\link{FALSE}}, the centering step is skipped.}

\item{cov}{covariance matrix (\eqn{p \times p}{p x p}) of the distribution.}

\item{inverted}{logical.  If \code{TRUE}, \code{cov} is supposed to
    contain the \emph{inverse} of the covariance matrix.}

\item{...}{passed to \code{\link{solve}} for computing the inverse of
    the covariance matrix (if \code{inverted} is false).}
}
\description{
The Mahalanobis distance function included in the `stats` package returns a complex number when given complex values of `x`.
But a distance (and thus its square) is always positive real. This function calculates the Mahalanobis distance using
the conjugate transpose if given complex data, otherwise it calls [stats::mahalanobis()].
}
\examples{
set.seed(4242)
n <- 8
x <- matrix(complex(real = rnorm(n), imaginary = rnorm(n)), ncol = 2)
mu <- complex(real = 1.4, imaginary = 0.4)
sigma <- 3.4
mahalanobis(x, mu, sigma * diag(2))
}
\references{
D. Dai and Y. Liang, High-Dimensional Mahalanobis Distances of Complex Random Vectors, Mathematics 9, 1877 (2021).
}
