% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{median}
\alias{median}
\title{Median, for Numeric or Complex}
\usage{
median(x, na.rm = FALSE, tol = 1e-07, maxiter = 200)
}
\arguments{
\item{x}{a numeric or complex vector, of which the median is to be calculated.}

\item{na.rm}{logical. Should NA values be removed before finding the median?}

\item{tol}{numeric. Relative tolerance to be passed to [pracma::geo_median]. Default is 1e-07.}

\item{maxiter}{maximum number of iterations for calculating geometric median. Not used if x is numeric.}
}
\value{
The median of x. If x is complex, the geometric median as calculated by Weiszfeld's algorithm.
}
\description{
Extends [stats::median] to understand complex variable input. If x is complex, the geometric median
is calculated by [pracma::geo_median], and returned as a complex number. Otherwise, [stats::median] is called.
}
\examples{
foo <- complex(real = 1:5, imaginary = 1:5)
median(foo)
}
\seealso{
[stats::median] and [pracma::geo_median]
}
