% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{pseuzvar}
\alias{pseuzvar}
\title{Pseudo-Variance of Complex Variables}
\usage{
pseuzvar(x)
}
\arguments{
\item{x}{a vector of complex numbers.}
}
\value{
complex. The pseudo-variance of \code{x}. If \code{x} is numeric, this is just the variance.
}
\description{
Calculates the pseudo-variance, also called the relational variance of a vector of complex numbers.
This describes the degree of covariance between the real and imaginary components.
}
\examples{
n = 6
z <- complex(real = rnorm(n), imaginary = rnorm(n))
pseuzvar(z)
}
