% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{var}
\alias{var}
\title{Variance for Complex Data}
\usage{
var(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{a numeric or complex vector, matrix, or dataframe.}

\item{y}{NULL (default) or a numeric vector, matrix, or dataframe with dimensions compatible with x. For complex x,
this parameter is not used as \code{cov()} and \code{cor()} are not implemented for complex variables.}

\item{na.rm}{logical. Should missing values be removed?}

\item{use}{character string giving the desired method of computing covariances in the presense of missing values. Not used.}
}
\value{
numeric, the sample variance of the input data.
}
\description{
A wrapper for \link[stats:cor]{stats::var} that can handle complex variables.
}
\details{
The sample variance is calculated as \code{sum(Conj( mean(x) - x ) * ( mean(x) - x )) / (length(x) - 1)}.
}
\examples{
set.seed(4242)
n <- 9
foo <- complex(real = rnorm(n), imaginary = rnorm(n))
var(foo)
}
