% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexlm-package.R
\docType{package}
\name{complexlm-package}
\alias{complexlm}
\alias{complexlm-package}
\title{complexlm: Linear fitting for complex valued data}
\description{
Tools for linear fitting with complex variables. Includes ordinary least-squares (zlm) and robust M-estimation (rzlm). Originally adapted from the rlm functions of MASS and the lm functions of stats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/QuantumOfMoose/complexlm}
}

}
\author{
\strong{Maintainer}: William Ryan \email{wryan1@binghamton.edu} [copyright holder]

Other contributors:
\itemize{
  \item Bruce White [thesis advisor]
  \item Brian Ripley (Wrote MASS rlm code upon which complexlm rlm is based) [contributor, copyright holder]
  \item Bill Venables [contributor]
  \item John Maindonald (Wrote plot.lm on which plot.zlm is based) [contributor]
  \item Martin Maechler (Wrote plot.lm on which plot.zlm is based) [contributor]
  \item R Core Team (Several functions in complexlm are based on similar functions in base or stats) [contributor, copyright holder]
}

}
\keyword{internal}
