% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zlm.wfit.R
\name{vcov.lm}
\alias{vcov.lm}
\title{Calculate Variance-Covariance Matrix and Pseudo Variance-Covariance Matrix for a Complex Fitted Model Object

A version of [stats::vcov] that is compatible with complex linear models. In addition to variance-covariance matrix,
the pseudo variance-covariance matrix, which is a measure of the covariance between real and imaginary components, is returned as well.}
\usage{
\method{vcov}{lm}(object, ...)
}
\arguments{
\item{object}{a fitted model object, typically. Sometimes also a summary() object of such a fitted model.}

\item{...}{Additional parameters, not currently used for anything.}
}
\value{
A list containing both the numeric variance-covariance matrix, and the complex pseudo variance-covariance matrix.
}
\description{
Calculate Variance-Covariance Matrix and Pseudo Variance-Covariance Matrix for a Complex Fitted Model Object

A version of [stats::vcov] that is compatible with complex linear models. In addition to variance-covariance matrix,
the pseudo variance-covariance matrix, which is a measure of the covariance between real and imaginary components, is returned as well.
}
\examples{
set.seed(4242)
n = 8
slop = complex(real = 4.23, imaginary = 2.323)
interc = complex(real = 1.4, imaginary = 1.804)
tframe <- data.frame(x= x <- complex(real= rnorm(n), imaginary= rnorm(n)), y= slop*x + interc)
fit <- lm(y ~ x, data = tframe, weights = rep(1,n))
vcov.lm(fit)
}
