% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlm.R
\name{vcov.rlm}
\alias{vcov.rlm}
\title{Calculate Variance-Covariance Matrix and Pseudo Variance-Covariance Matrix for a Complex Fitted Model Object}
\usage{
\method{vcov}{rlm}(object, ...)
}
\arguments{
\item{object}{a fitted model object, typically. Sometimes also a summary() object of such a fitted model.}

\item{...}{Additional parameters, not currently used for anything.}
}
\value{
A list containing both the numeric variance-covariance matrix, and the complex pseudo variance-covariance matrix.
}
\description{
A version of [stats::vcov] that is compatible with complex linear models. In addition to variance-covariance matrix,
the pseudo variance-covariance matrix, which is a measure of the covariance between real and imaginary components, is returned as well.
}
\examples{
set.seed(4242)
n = 8
slope = complex(real = 4.23, imaginary = 2.323)
intercept = complex(real = 1.4, imaginary = 1.804)
e <- complex(real=rnorm(n)/6, imaginary=rnorm(n)/6)
x <- complex(real = rnorm(n), imaginary = rnorm(n))
y <- slope * x + intercept + e
robfit <- rlm(x = x, y = y, weights = rep(1,n), interc = TRUE)
vcov.rlm(robfit)
}
